==> 1_environment_setup.ipynb <==
## Setup Environment
Use your own python
I used 3.9.6

# Saves the installed requirements to a requirements file
%pip freeze > requirements.txt

# Installs the requirements from the requirements file
%pip install -r requirements.txt
