==> 3_1_sentiment_analysis_feature.ipynb <==
# Read Data

# Import Libraries and Data
from transformers import pipeline
import pandas as pd
import warnings 

warnings.simplefilter(action='ignore', category=FutureWarning)

df = pd.read_csv("source data/twitter_human_bots_dataset.csv", index_col=0)
df.head()
df_sentiment = df[['screen_name', 'description', 'account_type']]

## Sentiment Analysis for `description` feature

# Preprocess the description column
description = df_sentiment['description'].fillna('').str.lower()
description = description.str.replace(r'@\w+', '', regex=True)
description = description.str.replace(r'http\S+', '', regex=True)
description = description.str.replace('#', '', regex=True)
description = description.str.replace(r'[^a-zA-Z0-9\s]', '', regex=True)

# Get description statistics
print(len(description))
description_lengths = description.apply(len)
print(description_lengths.describe())
print(description)

# Initialise model
model_name = 'bert-base-uncased'
sentiment_model = pipeline('sentiment-analysis')

# Get sentiment of the description
sentiments = description.apply(lambda text: sentiment_model(text)[0])

df1 = pd.DataFrame(sentiments.tolist())
print(df1.head())
# Takes 25 minutes to run


### Store into original dataframe

# Add sentiment label to original df
df_sentiment['sentiment_label'] = df1['label'].map({'NEGATIVE': 0, 'POSITIVE': 1})

df_sentiment.groupby('account_type')['sentiment_label'].value_counts()

### Save into external csv file

# Select relevant columns
df_sentiment = df_sentiment[['screen_name', 'sentiment_label']]

df_sentiment.to_csv('cleaned data/df_sentiment.csv', index=False)
