==> 6_results.ipynb <==
# Results

import pandas as pd

model_results = pd.read_csv('output_files/model_results.csv')
nn_results = pd.read_csv('output_files/model_results_nn.csv')
results_table = []
model_results.head()

print(model_results.loc[model_results["model"]=="RandomForest"]["metrics"])

# Model name, accuracy, precision, recall, f1, auc

LR_log = {"Model": "Log Regression with Log features", "Threshold": pd.NA, "Precision": 0.82, "Recall": 0.82, "F1": 0.82 , "AUC": pd.NA}
LR_nonlog = {"Model": "Log Regression with non-Log features", "Threshold": pd.NA, "Precision": 0.73, "Recall": 0.74, "F1": 0.72 , "AUC": pd.NA}

RF_050 = {"Model": "Random Forest", "Threshold": 0.5, "Precision": 0.853, "Recall": 0.738, "F1": 0.791 , "AUC": 0.840}
RF_080 = {"Model": "Random Forest", "Threshold": 0.8, "Precision": 0.853, "Recall": 0.738, "F1": 0.791 , "AUC": 0.840}
RF_085 = {"Model": "Random Forest", "Threshold": 0.85, "Precision": 0.853, "Recall": 0.738, "F1": 0.791 , "AUC": 0.840}
RF_090 = {"Model": "Random Forest", "Threshold": 0.9, "Precision": 0.853, "Recall": 0.738, "F1": 0.791 , "AUC": 0.840}
RF_095 = {"Model": "Random Forest", "Threshold": 0.95, "Precision": 0.853, "Recall": 0.738, "F1": 0.791 , "AUC": 0.840}

XG_050 = {"Model": "XGBoost", "Threshold": 0.5, "Precision": 0.844, "Recall": 0.779, "F1": 0.810 , "AUC": 0.857}
XG_080 = {"Model": "XGBoost", "Threshold": 0.8, "Precision": 0.949, "Recall": 0.594, "F1": 0.731 , "AUC": 0.790}
XG_085 = {"Model": "XGBoost", "Threshold": 0.85, "Precision": 0.962, "Recall": 0.547, "F1": 0.697 , "AUC": 0.768}
XG_090 = {"Model": "XGBoost", "Threshold": 0.9, "Precision": 0.977, "Recall": 0.474, "F1": 0.638 , "AUC": 0.735}
XG_095 = {"Model": "XGBoost", "Threshold": 0.95, "Precision": 0.992, "Recall": 0.357, "F1": 0.526 , "AUC": 0.678}

LG_050 = {"Model": "LGBM", "Threshold": 0.5, "Precision": 0.843, "Recall": 0.775, "F1": 0.808 , "AUC": 0.855}
LG_080 = {"Model": "LGBM", "Threshold": 0.8, "Precision": 0.955, "Recall": 0.596, "F1": 0.734 , "AUC": 0.792}
LG_085 = {"Model": "LGBM", "Threshold": 0.85, "Precision": 0.964, "Recall": 0.534, "F1": 0.688 , "AUC": 0.763}
LG_090 = {"Model": "LGBM", "Threshold": 0.9, "Precision": 0.977, "Recall": 0.454, "F1": 0.620 , "AUC": 0.725}
LG_095 = {"Model": "LGBM", "Threshold": 0.95, "Precision": 0.994, "Recall": 0.330, "F1": 0.496 , "AUC": 0.665} 

nn = {"Model": "NN", "Threshold": pd.NA, "Precision": 0.795, "Recall": 0.601, "F1": 0.684 , "AUC": pd.NA}

AS = {"Model": "Sentiment Analysis of User's Description", "Threshold": pd.NA, "Precision": 0.594, "Recall": 0.531, "F1": 0.560 , "AUC": pd.NA}

data = [LR_log, LR_nonlog, LG_050, LG_080, LG_085, LG_090, LG_095, RF_050, RF_080, RF_085, RF_090, RF_095, XG_050, XG_080, XG_085, XG_090, XG_095, nn, AS]
df = pd.DataFrame(data)

print(df)
